Compile PPTP as a module
I tried everything however I still can't get it to work. So I will just
compile it as a module, if someone wants to use they can. PPTP is a very
insecure protocol btw, so you should not use it (SSH is a nice alternative
that works with default image).