Added "m45t3r" profile and misc files
"m45t3r" target is a profile based on "default" that includes extra plugins
and reactivates miniupnpd. It is a profile with the objective to extract the
maximum from cheap routers with 4MB of Flash and no USB support.

Re-enable CONFIG_USE_MKLIBS. Seems silly to save space when you can not
install most external packages, however I am trying to fix one strange issue
that sometimes the Wi-Fi simple stops working (and the only fix is a Factory
Reset, not even reboot works). Should save an additional 64KiB.

Adds some extra devices that is not included in official builds, that is:
TL-WR740n-v5, TL-WR740n-v6 and TL-WR741nd-v5. Only TL-WR740n-v5 was tested
and seems to work fine. No support here, you can try if you want.

Makefile/build.sh is modified to point to upstream commit id instead of the
current one from this repository, to make it easier to know which upstream
version this firmware is based on.

create_archive_for_m45t3r_builds.sh is just a small script that creates a 7z
file after compilation, including built images and packages to redistribution.

This is the main commit from this repository, and the idea is that every new
change in upstream needs only to be rebased from this commit. This is why it
includes all modified files in this repo.